+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.8.0
+
+- refactor: remove pages & components view
+- refactor: move scripts&dependencies view to explorer sidebar
+
+
 ## 0.7.6
 
 - fix: terminal names are the same when running npm scripts